--- lerobot/common/teleoperators/spes_teleop/cert.pem
@@ -1,31 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFazCCA1OgAwIBAgIUCqlPRtb5LcN3sZJQLecTgbCCPHYwDQYJKoZIhvcNAQEL
-BQAwRTELMAkGA1UEBhMCUlMxEzARBgNVBAgMClNvbWUtU3RhdGUxITAfBgNVBAoM
-GEludGVybmV0IFdpZGdpdHMgUHR5IEx0ZDAeFw0yNDA3MjgxNDU0NDZaFw0yNTA3
-MjgxNDU0NDZaMEUxCzAJBgNVBAYTAlJTMRMwEQYDVQQIDApTb21lLVN0YXRlMSEw
-HwYDVQQKDBhJbnRlcm5ldCBXaWRnaXRzIFB0eSBMdGQwggIiMA0GCSqGSIb3DQEB
-AQUAA4ICDwAwggIKAoICAQDkHaxyFbvoQB0QqVLMYAnhTaZ1lxj/6+lPpPQJJ8DH
-9qcpJb04ki27hZwCUBaHMg+8QX18JJa1M+jzRk7mEJl5tjcyYyD8iDxfUfhRzrB1
-l16E23yFKl06JieSNPBgUAXSCewoR4Qde9mIi9NaMGf1njaZ5VMjDpMHHCGgdSUI
-VGq6xoDJD63dizT+x4gqCStfbvCO4yG0KyStUh+GX9PcMpdizx7EtE85YuXBD+XF
-Nmm5gWlqPnHsWQwhwxi3XkQWMWT1OU0M6PIw6fjboSehC1s5DdkAI4dSY6vTbNLf
-QxSA8jk/rXq+SR3boPW39AK+U95mqmO6ZkUE05T3z112al1UkGd9TklTPDLLO/i9
-tiuPUly9349yYBS+5+lUn1Kc/vOF1/YzakG1KLqJbzQQJXv9+SeTQSLkdYHvm74y
-Z8Li+R23oavIJ3kZvfhKFVc64HWhijejXifmHRbuqss7hn4etHfGaawJTOGQWOvL
-vp+DOw80MJ+BRvVyY3Nl/XV1NszlX5o8r5jyo3ZJfHY65AjpdN9SVSyA1KxurSpL
-AkvjJoh73MwrVzl6awOPzzPV8gqrfxCSEKXsYOboOOAeONAm3R8nJksmAEJWz2v/
-70JcfOZR1RVMmLDbaX1fN8Bs6/l6vpESTctCmM7/yUYmjQOGx3lAZRPdjc5Uku55
-LQIDAQABo1MwUTAdBgNVHQ4EFgQU5zC52tTr26VvS7Z0Yz2fk6FyDUAwHwYDVR0j
-BBgwFoAU5zC52tTr26VvS7Z0Yz2fk6FyDUAwDwYDVR0TAQH/BAUwAwEB/zANBgkq
-hkiG9w0BAQsFAAOCAgEAQBEtymNkYw930Mrnjk82eBWhCci7sXowr4U9+aYV+P0u
-gvYf9THI8Xqtd68g/l/FsxgZmG4xaKHcrrlUSYQjljpFHs+UYY/I68hxeRVXs490
-OQybakwDVe+IXoMa7F1nPmjP5h5HkL4bEIH2XV7BS0MThBVBm2MPrHN2xPk8Ut4k
-BV+Edq1z73LgLwuPncihn9HZBY8OQUnYWTTAU0XmQfU+pRbZ8zqsvfaLs2gUEU0s
-zyBoOHKzHGJmg4WH9qxN2YwL6Cu1QbWeobB9YanbuUjd0z/uL7fNQADMMeIOkUN5
-homxxMr1f9TNVsYLPWYIsWVyy2JUnybLYnwYPQBCahbF5HU+avTOnTYkyUnevoVC
-1YV615LagBg98pi8PkPYd1QE9tP+sWNCZLAMWiG9NKIdwpJ3cAcqYGMvhewzjpH6
-JVsHP4Xpic8urlQFsoRMc42Ldq/kI3+5PqW9wVfmKjX4WthbvpuUzbmy70bvdokM
-PzYGsdYSAcZGU/pV/8b/PtDhlYGWpucAdZv718VHJ28Z9e320Zs1ZBe6QNvD4Pg4
-AI7V6rIv/SOT7W8tuyG/zMtCokbcNoaeMhHD67d9gB2fgAs7p7qaChvdb3C9o1Pz
-jIC0+EgNx/ukv57IJlOR+Kfx+mtia92v1MiWZTxarIWfPwKLZziVBnouldZCb4A=
------END CERTIFICATE-----
