--- lerobot/common/teleoperators/spes_teleop/key.pem
@@ -1,52 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIJRAIBADANBgkqhkiG9w0BAQEFAASCCS4wggkqAgEAAoICAQDkHaxyFbvoQB0Q
-qVLMYAnhTaZ1lxj/6+lPpPQJJ8DH9qcpJb04ki27hZwCUBaHMg+8QX18JJa1M+jz
-Rk7mEJl5tjcyYyD8iDxfUfhRzrB1l16E23yFKl06JieSNPBgUAXSCewoR4Qde9mI
-i9NaMGf1njaZ5VMjDpMHHCGgdSUIVGq6xoDJD63dizT+x4gqCStfbvCO4yG0KySt
-Uh+GX9PcMpdizx7EtE85YuXBD+XFNmm5gWlqPnHsWQwhwxi3XkQWMWT1OU0M6PIw
-6fjboSehC1s5DdkAI4dSY6vTbNLfQxSA8jk/rXq+SR3boPW39AK+U95mqmO6ZkUE
-05T3z112al1UkGd9TklTPDLLO/i9tiuPUly9349yYBS+5+lUn1Kc/vOF1/YzakG1
-KLqJbzQQJXv9+SeTQSLkdYHvm74yZ8Li+R23oavIJ3kZvfhKFVc64HWhijejXifm
-HRbuqss7hn4etHfGaawJTOGQWOvLvp+DOw80MJ+BRvVyY3Nl/XV1NszlX5o8r5jy
-o3ZJfHY65AjpdN9SVSyA1KxurSpLAkvjJoh73MwrVzl6awOPzzPV8gqrfxCSEKXs
-YOboOOAeONAm3R8nJksmAEJWz2v/70JcfOZR1RVMmLDbaX1fN8Bs6/l6vpESTctC
-mM7/yUYmjQOGx3lAZRPdjc5Uku55LQIDAQABAoICABPeo9EyJHJ39zAm14/2w5WM
-6nVuql5MYb+uuhT8EfY4/Vl/sIo3bWVan3qyGCxBh45Z1ADWvQvekCl0k9XFV2D7
-JrasXMiyYzPYU31ouAUeIL0I00FULdetUCATnWZ+QeuIMrpVdGQPECUtLMU71ypV
-VCM2jvXMRLFHWYR8/7NIix1mwfwBch8Mi5qvraQftVf0NV702NA1O9YfA6D0QuN4
-wqf4K5FttULl4uZcqKJzeWUPqSt4cfrozL46HLKIMv0N42Ha0mPP/JyXHLagxG9q
-KlF/a8nvzk82YFajsYCB+sy6jJMH3NurSb8CN7UEu6JxcrtVlyc61kMz3NNzDAxk
-3qg/PhCvtnAdxj1c6WWH8kp/J0uzaF/VdQIsmwl/MNOCSbNq4hDXTjDuu3wW5W1G
-fiVkbJe48IgEgV37xJ80NrJeeXv9V1Mj/vXzW238QOjiD1zsR3a09otrlAtuTiUN
-kHdbTdCbqNcSlgDsCLq3pcSVlYbomDiMOSwt6XLWV7A87zavAPXuEFi5EWGAhqAM
-XBIAwmfw86H2bM/vpOspzkuTPXrjAC+LwT03vsUhFgeZ9qptDE7KoYtYKwCSP9ob
-wUX7H+jUmpxHIcSVpfUI/IRyZJIAOOL93Zu8Xo8IWEckxpupoBu0mF6iVG3BKyGy
-dEg8ko0pRRmZ67had9cRAoIBAQD233C3gPV8KGJryLdzbJ4lW2xZEZ3xgNHBotxl
-8NwtW5B33NN48mv+ELCsVPiFswX1n8YrA5w1lt/6u9qM0b3pE6U5KfQQaged5Wjj
-mfIAafUsAGB1W6m8ObYn/q6saUGncSgynaECWb1N7hCuOEIyj3+L7bpP1CNS94mM
-8DToJMlAKSYeYaXsA91oZEC1XIe4F94liQ6sFktgyywXGymeHnYaMa5Cwew7WZeQ
-aaF/9bbOoER0fK+QB+bvetaBPtSXiJjXAzUn5iA1B/u4YG+5wte3mPxJO0yf7mMq
-xJfXb3w0mXGQ64z5SG0Zk+h8fprAf8EI5XlVBPY8r1NCSNLFAoIBAQDsjLTJWPqu
-C4MimZt01O6w8+dDvwMElq1zyBJgA10DEf34YZRSKtRMjTWBYNfF8OBlBymhhR2z
-TV7Rqrogz+9EO6YB+KGa1vQeH253sQvKn2oe6PTFSUOV/YHcrY/1s6VZyqmbkDMx
-2Hx/4Ne4FLS9H8RkNiZwEdXOEK6DQFDpEvA8LLMdRV9yMq3jalqrbNFJOGY92fGf
-SONQ0idG+xIchC9LY6VJXCHT1w2GZ5M8J2BpQBiHpBqEXoVx2uTWk6k2kz5FWJnI
-Pxa2TjZuwssdrs+k5oflAl+jWNqughgf7TogNmCST/afd4AOu7nta7ssmNGc0XO0
-AFWody5rytNJAoIBAQCSRvLSyGasmoE2vEd27PUHzZP1cqrTXjDrO0S8YmsTtR10
-6DOlhn7Wm60Tb4WSMvFlCLay5gC5ABzWxFhK+Dw+zpUnueQTlb/NnxD3S1BJTn8n
-A1IkN5vrG4cDS7TUE8oDHNlkpGGystmoQ1c+47bVnpFEqCetBVWo42NtTycKJkHY
-8ojTl6hooCGgqUZuL9V7DylhtEQtU0h2oewOBNo0BniULLvEcoM2528W8n/vAfi/
-ldb57qdYAFalr3NKQaBV6WomfJAShBixU3mjVgDdYMXQ06Uo/yn32ZiIztPpovVF
-dWYFKIzvaCq87RbkLWAkj44hZ1btOXNtobUDoYLBAoIBAQDQjhnSr72Md5/tgIq8
-DgRIRTJv21fY0tQokYwJ1M1LqLBHLhkr/JDi0VkMMYrr/LyKti7xNkECVcO0jZa+
-2TV17jUl5cTtlHj0mfe2VYU05yoUCIXN7xNEiZbWLrwO/41YRkMVYaSRXPIfeOgY
-Z0EWwYxQYiGvizn7DWAfXjQ32jeuhiaBNym4sI/5QkqgezdmG/Cgtae/LO8l0bNJ
-dJD6TVy5VihjLnSwphlJl8YLMhhhk15NMNUkQYXb5Qb/XH7DHDy5vS0QjPcEUp7R
-NrMH/5kZJRtJQLd7eBqN4ytL8sbdobOxLAkffaeM1dxORD5pSm78mPovCGB9QF7Q
-MKnpAoIBAQDN4RrdxFA2dVB3ooqjjbCfUOfMyFG9SeMzqzWIR8LOT3FYW8bM2tUU
-kXSxzcJ1ep4tX9dVKKI7WhYs3IpZMQFnUE/1+MXZ7fK0Za03s3wixsZaSX25el1/
-9mQGu8Rdecpjv0H0KHUHgNXbGoU1MYEu6NH1movs7Bd86iN9yAI88FmSci/Xmogi
-FKiQUqXHcZsOZmIBmqhgiFrdthFnMvin/giBkvTYS3M8POTz54GB4Z+Y4J6tkdxW
-d5TGDT+Yuk38frH/Jhw5cNImZdvHXEveIbjsFkx3W9ccAwnRKIyE9Kps0NXGiApW
-woxgMahddE/gJ4Bq4NQ714vw5oG542n3
------END PRIVATE KEY-----
