+++ lerobot/common/teleoperators/telephone/__init__.py
@@ -1,6 +1,6 @@
-#!/usr/bin/env python
+# !/usr/bin/env python
 
-# Copyright 2024 The HuggingFace Inc. team. All rights reserved.
+# Copyright 2025 The HuggingFace Inc. team. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,14 +14,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from dataclasses import dataclass
-
-from ..config import TeleoperatorConfig
-
-
-@TeleoperatorConfig.register_subclass("spes_teleop")
-@dataclass
-class SpesTeleopConfig(TeleoperatorConfig):
-    port: str = "4443"
-    host: str = "0.0.0.0"
-    use_gripper: bool = True
+from .configuration_telephone import TelephoneConfig
+from .telephone import Telephone
