+++ lerobot/record.py
@@ -88,7 +88,7 @@ from lerobot.common.utils.visualization_utils import _init_rerun
 from lerobot.configs import parser
 from lerobot.configs.policies import PreTrainedConfig
 
-from .common.teleoperators import koch_leader, so100_leader, so101_leader, spacemouse, spes_teleop  # noqa: F401
+from .common.teleoperators import koch_leader, so100_leader, so101_leader, spacemouse, telephone  # noqa: F401
 
 
 @dataclass
